sdk/uis/tools/openapi_import/requirements.txt
pyyaml>=5.4.0
requests>=2.25.0
